refactor(container.js): fixed broken auto loadead dependencies
