+++ src/interfaces/http/graphQL/resolvers/queries/index.js
@@ -1,4 +1,4 @@
-const heroQueries = require('./heroQuerys');
+const heroQueries = require('./heroQueries');
 
 module.exports = {
   heroQueries,
